Move PlainTransactionInfoResponse out of messages_grpc.rs
