+++ crates/sui-core/src/quorum_driver/mod.rs
@@ -33,7 +33,7 @@ use mysten_common::sync::notify_read::{NotifyRead, Registration};
 use mysten_metrics::{spawn_monitored_task, GaugeGuard};
 use std::fmt::Write;
 use sui_types::error::{SuiError, SuiResult};
-use sui_types::messages_grpc::PlainTransactionInfoResponse;
+use sui_types::messages_safe_client::PlainTransactionInfoResponse;
 use sui_types::transaction::{Transaction, VerifiedCertificate};
 
 use self::reconfig_observer::ReconfigObserver;
