+++ crates/sui-types/src/messages_grpc.rs
@@ -7,7 +7,7 @@ use crate::effects::{
     SignedTransactionEffects, TransactionEvents, VerifiedSignedTransactionEffects,
 };
 use crate::object::{Object, ObjectFormatOptions};
-use crate::transaction::{SenderSignedData, SignedTransaction, Transaction, VerifiedCertificate};
+use crate::transaction::{SenderSignedData, SignedTransaction};
 use move_core_types::value::MoveStructLayout;
 use serde::{Deserialize, Serialize};
 
@@ -147,33 +147,6 @@ pub struct HandleTransactionResponse {
 pub struct TransactionInfoResponse {
     pub transaction: SenderSignedData,
     pub status: TransactionStatus,
-}
-
-/// This enum represents all possible states of a response returned from
-/// the safe client. Note that [struct SignedTransaction] and
-/// [struct SignedTransactionEffects] are represented as an Envelope
-/// instead of an VerifiedEnvelope. This is because the verification is
-/// now performed by the authority aggregator as an aggregated signature,
-/// instead of in SafeClient.
-#[derive(Clone, Debug)]
-pub enum PlainTransactionInfoResponse {
-    Signed(SignedTransaction),
-    ExecutedWithCert(
-        VerifiedCertificate,
-        SignedTransactionEffects,
-        TransactionEvents,
-    ),
-    ExecutedWithoutCert(Transaction, SignedTransactionEffects, TransactionEvents),
-}
-
-impl PlainTransactionInfoResponse {
-    pub fn is_executed(&self) -> bool {
-        match self {
-            PlainTransactionInfoResponse::Signed(_) => false,
-            PlainTransactionInfoResponse::ExecutedWithCert(_, _, _)
-            | PlainTransactionInfoResponse::ExecutedWithoutCert(_, _, _) => true,
-        }
-    }
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
